added a container so that the icons don't wrap in Firefox, which it does for some reason. Behavior in chrome seems to be unchanged.
