+++ web/src/components/Home/Home.tsx
@@ -1,19 +1,14 @@
 import moment from 'moment-timezone';
 import * as React from 'react';
-import { Transition } from 'react-transition-group';
 
 import { css } from '@emotion/core';
 import styled from '@emotion/styled';
 
-import { Elastic } from 'gsap/EasePack';
-import TweenLite from 'gsap/TweenLite';
-
-import { lato1, lato2, lato2i } from 'src/styles/fonts';
+import { lato1, lato2 } from 'src/styles/fonts';
 import {
     generateSrcsetWidths,
     homeBackground,
     resizedImage,
-    staticImage,
     sycChairVertical,
 } from 'src/styles/imageUrls';
 import { container, noHighlight } from 'src/styles/mixins';
@@ -22,6 +17,8 @@ import { navBarHeight } from 'src/styles/variables';
 import { DesktopBackgroundPreview, MobileBackgroundPreview } from 'src/components/Home/PreviewSVGs';
 import { LazyImage } from 'src/components/LazyImage';
 import { screenLengths, screenXSandPortrait, screenXSorPortrait } from 'src/styles/screens';
+import socials from 'src/components/Home/socials';
+import Social from 'src/components/Home/Social';
 
 const textShadowColor = 'rgba(0, 0, 0, 0.75)';
 
@@ -76,26 +73,6 @@ const Skills = styled('div')`
         font-size: calc(100vw / 16);
         bottom: 58%;
         top: unset;
-    }
-`;
-
-const Handle = styled('div')`
-    ${noHighlight}
-    margin: 15px 0;
-    font-family: ${lato2i};
-    font-size: 30px;
-    color: white;
-    text-shadow: 0 0 6px ${textShadowColor};
-    text-decoration: underline;
-    -webkit-tap-highlight-color: transparent;
-
-    &:hover {
-        cursor: pointer;
-    }
-
-    ${screenXSorPortrait} {
-        width: 100%;
-        bottom: 15%;
     }
 `;
 
@@ -190,20 +167,6 @@ const NavBarGradient = styled('div')`
     }
 `;
 
-const SocialContainer = styled('div')`
-    position: absolute;
-    width: fit-content;
-    top: 55%;
-    left: 50%;
-    transform: translateX(-50%);
-
-    ${screenXSandPortrait} {
-        bottom: 12%;
-        font-size: 0.8rem;
-        top: unset;
-    }
-`;
-
 const StyledCopyright = styled('div')`
     position: absolute;
     bottom: 0;
@@ -221,136 +184,6 @@ const StyledCopyright = styled('div')`
 const srcWidths = screenLengths.map((value) => (
     Math.round(value * 1779 / 2560)
 ));
-
-const socials: {
-    [key: string]: string;
-} = {
-    facebook: 'https://www.facebook.com/seanchenpiano',
-    twitter: 'https://twitter.com/seanchenpiano',
-    youtube: 'https://www.youtube.com/user/SeanChenPiano',
-    spotify: 'https://open.spotify.com/artist/6kMZjx0C2LY2v2fUsaN27y?si=8Uxb9kFTQPisQCvAyOybMQ',
-    linkedin: 'https://www.linkedin.com/in/seanchenpiano',
-    instagram: 'https://www.instagram.com/seanchenpiano',
-};
-
-const SocialLink = styled.a<{ show: boolean; canHover: boolean; }>`
-    padding: 1.5rem 0;
-    width: calc(100vw / ${Object.keys(socials).length});
-    max-width: 120px;
-    display: inline-block;
-    height: 100%;
-    opacity: 0;
-    pointer-events: ${props => props.show ? 'unset' : 'none'};
-    filter: drop-shadow(0 0 0.5rem black);
-
-    ${screenXSandPortrait} {
-        padding: 0.8rem 0;
-    }
-
-    ${props => props.canHover ? `
-        transition: transform 0.1s linear, filter 0.1s linear;
-        &:hover {
-            transform: scale(1.1);
-            filter: drop-shadow(0 0 0.75rem black);
-        }
-    ` : ''}
-`;
-
-interface SocialMediaLinkProps {
-    className?: string;
-    social: string;
-    show: boolean;
-    canHover: boolean;
-    url: string;
-}
-
-const SocialMediaLink: React.FC<SocialMediaLinkProps> = (props) => (
-    <SocialLink canHover={props.canHover} show={props.show} href={props.url} target="_blank">
-        <img
-            className={''}
-            src={staticImage(`/logos/${props.social}.svg`)}
-        />
-    </SocialLink>
-);
-
-interface SocialState {
-    show: boolean;
-    canHover: { [key: string]: boolean; };
-}
-
-class Social extends React.PureComponent<{}, SocialState> {
-    defaultCanHover = Object.keys(socials).reduce((prev, curr) => {
-        return {
-            ...prev,
-            [curr]: false,
-        };
-    }, {});
-
-    state: SocialState = {
-        show: false,
-        canHover: this.defaultCanHover,
-    };
-
-    onHandleClick = () => {
-        this.setState({ show: !this.state.show });
-    }
-
-    onSocialEnter = (id: number) => (el: HTMLLinkElement) => {
-        const relative = id - (Object.keys(socials).length / 2 - 0.5);
-        TweenLite.fromTo(el, 0.25,
-            {
-                opacity: 0,
-                y: `-50%`,
-                x: `${relative * -100}%`,
-            }, {
-                opacity: 1,
-                y: `0%`,
-                x: `0%`,
-                delay: .05 * id,
-                ease: Elastic.easeOut.config(1, 0.75),
-                clearProps: 'transform',
-            });
-    }
-
-    onSocialExit = (id: number) => (el: HTMLLinkElement) => {
-        const relative = id - Math.floor(Object.keys(socials).length / 2);
-        TweenLite.fromTo(el, 0.25,
-            {
-                opacity: 1,
-                y: `0%`,
-                x: `0%`,
-            }, {
-                opacity: 0,
-                y: `-50%`,
-                x: `${relative * -100}%`,
-                delay: .05 * id,
-                ease: Elastic.easeOut.config(1, 0.75),
-                clearProps: 'transform',
-            });
-        this.setState({ canHover: this.defaultCanHover });
-    }
-    render() {
-        return (
-            <SocialContainer>
-                <Handle onClick={this.onHandleClick}>@seanchenpiano</Handle>
-                {
-                    Object.keys(socials).map((key, idx) => (
-                        <Transition
-                            key={key}
-                            in={this.state.show}
-                            onEnter={this.onSocialEnter(idx)}
-                            onExit={this.onSocialExit(idx)}
-                            timeout={250 + 50 * idx}
-                            onEntered={() => this.setState({ canHover: { ...this.state.canHover, [key]: true } })}
-                        >
-                            <SocialMediaLink canHover={this.state.canHover[key]} show={this.state.show} url={socials[key]} social={key} />
-                        </Transition>
-                    ))
-                }
-            </SocialContainer>
-        );
-    }
-}
 
 interface HomeProps {
     bgLoaded: () => void;
