+++ web/src/components/Home/socials.ts
@@ -1,0 +1,12 @@
+const socials: {
+    [key: string]: string;
+} = {
+    facebook: 'https://www.facebook.com/seanchenpiano',
+    twitter: 'https://twitter.com/seanchenpiano',
+    youtube: 'https://www.youtube.com/user/SeanChenPiano',
+    spotify: 'https://open.spotify.com/artist/6kMZjx0C2LY2v2fUsaN27y?si=8Uxb9kFTQPisQCvAyOybMQ',
+    linkedin: 'https://www.linkedin.com/in/seanchenpiano',
+    instagram: 'https://www.instagram.com/seanchenpiano',
+};
+
+export default socials;
